requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
clickhouse-connect==0.7.19
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
streamlit
prometheus-api-client
